fix: error.mp4 not showing + unsafe file name

+++ streamer.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
 stream() {
+    echo "streaming ${1}"
     ffmpeg -hide_banner -loglevel error -re -i ${1} -c copy -f flv rtmp://a.rtmp.youtube.com/live2/
+    echo "stream ended"
 }
 
 if test -f buffer/playlist.mp4; then
@@ -13,7 +15,7 @@ else
 fi
 
 while(true); do
-    if [ ${count} -eq 0 ]; then
+    if test -f buffer/playlist.mp4 && [ ${count} -eq 0 ]; then
         mv buffer/playlist.mp4 playlist.mp4
 
         ./renderer.sh &
@@ -21,6 +23,11 @@ while(true); do
          
         stream playlist.mp4
     else
+        if [ ${count} -eq 0 ]; then
+            ./renderer.sh &
+            pid=$!
+        fi
+
         cp buffer/video_buffer/error.mp4 error.mp4
         stream error.mp4
     fi
